change to use bigint

+++ migrations/20170625141346-create-channels.js
@@ -3,7 +3,7 @@
 module.exports = {
   up: function (queryInterface, Sequelize) {
     return queryInterface.createTable('channels', {
-      id: { type: Sequelize.INTEGER, primaryKey: true },
+      id: { type: Sequelize.BIGINT, primaryKey: true },
       created_at: { type: Sequelize.DATE, defaultValue: Sequelize.literal('CURRENT_TIMESTAMP'), allowNull: false },
       updated_at: { type: Sequelize.DATE },
       deleted_at: { type: Sequelize.DATE },
