+++ src/models/entity/channel.ts
@@ -18,7 +18,7 @@ export interface ChannelModel extends Sequelize.Model<ChannelInstance, ChannelAt
 
 export function defineChannel<ChannelInstance, ChannelAttribute>(sequelize: SequelizeInterface) {
   const Channel = sequelize.define('Channel', {
-    id: { type: Sequelize.INTEGER, primaryKey: true },
+    id: { type: Sequelize.BIGINT, primaryKey: true },
     created_at: { type: Sequelize.DATE, defaultValue: Sequelize.literal('CURRENT_TIMESTAMP'), allowNull: false },
     updated_at: { type: Sequelize.DATE },
     deleted_at: { type: Sequelize.DATE },
