+++ src/models/entity/group-user.ts
@@ -25,7 +25,7 @@ export function defineGroupUser<GroupUserInstance, GroupUserAttribute>(sequelize
     updated_at: { type: Sequelize.DATE },
     deleted_at: { type: Sequelize.DATE },
     group_id: { type: Sequelize.INTEGER, allowNull: false },
-    user_id: { type: Sequelize.INTEGER, allowNull: false }
+    user_id: { type: Sequelize.BIGINT, allowNull: false }
   }, {
     tableName: 'group_users',
     timestamps: true,
